Revert "use fp32 during all-reduce"
This reverts commit 164be9ff54c19288eabf57fecfa9a0d41daded7f.

+++ pytorch-fsdp/fsdp_finetune.py
@@ -93,7 +93,7 @@ class Trainer:
             model,
             mixed_precision=MixedPrecision(
                 param_dtype=torch.float16,
-                reduce_dtype=torch.float32,
+                reduce_dtype=torch.float16,
                 buffer_dtype=torch.float16,
             ),
             forward_prefetch=False,
